+++ package.json
@@ -24,6 +24,6 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "react-router-dom": "^6.16.0",
     "typescript": "^5.0.2",
-    "vite": "^4.4.5"
+    "vite": "^4.4.11"
   }
 }
