+++ src/components/layout/Header/Ingredients-content/Herbs/Rosemary/Rosemary_SideEffects.tsx
@@ -32,7 +32,7 @@ function Rosemary_SideEffects() {
             <div className="component-content-rendering">
                 <Outlet />
                 <div className="flex-content">
-                    <Ingredient ingredientName="Rosemary"
+                    <Ingredient ingredientName="Salvia rosmarinus(Rosemary)"
                         familyName="Lamiaceae" origin="Native to the Mediterranean region"
                         geoLocation="Mediterranean Region" />
                     <div className="content-wrapper">
