--- src/components/layout/Header/Ingredients-content/Plants/FlaxSeeds.tsx
@@ -1,329 +1,0 @@
-import "../../../../../App.css"
-import Ingredient from "../Ingredient";
-import IngredientNameHeading from "../Ingredient-heading";
-
-function Ingredient_Overview() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>What are Flax Seeds?</h2>
-                    <p>
-                        Flax, also known as common flax or linseed, is a flowering plant,
-                        Linum usitatissimum, in the family Linaceae. It is cultivated as a
-                        food and fiber crop in regions of the world with temperate climates.
-                    </p>
-                    <p>
-                        Flaxseed is a plant-based food that provides healthful fat,
-                        antioxidants, and fiber. It can provide many health benefits.
-                        People grew flax as a crop in ancient Egypt and China. In Asia,
-                        it has had a role in Ayurvedic medicine for thousands of years.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_HowToUse() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>How are Flax Seeds used?</h2>
-                    <p>
-                        Flaxseed are available in the form of:
-                    </p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>Seeds</li>
-                            <li>Oils</li>
-                            <li>Powder</li>
-                        </ul>
-                        <ul>
-                            <li>Tablets</li>
-                            <li>Capules</li>
-                            <li>Flour</li>
-                        </ul>
-                    </div>
-                    <p>Flax seeds are also used in:</p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>Cereals</li>
-                            <li>Yogurt</li>
-                            <li>Smoothies</li>
-                            <li>Chili</li>
-                        </ul>
-                        <ul>
-                            <li>Sauces</li>
-                            <li>Gravies</li>
-                            <li>Burritos</li>
-                            <li>Lasagna</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    People use it as a dietary supplement to prevent constipation, diabetes,
-                    high cholesterol, heart disease, cancer, and several other conditions.
-                    The nutrients in flaxseed include lignans, antioxidants, fiber, protein,
-                    and polyunsaturated fatty acids such as alpha-linolenic acid (ALA), or
-                    omega-3. Consuming these nutrients may help lower the risk of various
-                    conditions.
-                </p>
-                <p>
-                    However, there is not currently enough evidence to support all of these
-                    claims. Find out below, what the research says about flaxseed and its
-                    possible health benefits.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Benefits() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>What are the benefits of using Flax Seeds?</h2>
-                    <p>
-                        Like other plant-based foods, flaxseed is rich in antioxidants. These
-                        can help prevent disease by removing molecules called free radicals
-                        from the body.
-                    </p>
-                    <p>
-                        Free radicals occur as a result of natural processes and environmental
-                        pressures. If there are too many free radicals in the body, oxidative
-                        stress can develop, leading to cell damage and disease. Antioxidants
-                        help remove free radicals from the body.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h2>The health benefits of using Flax Seeds:</h2>
-                <h3>Reducing the risk of cancer:</h3>
-                <p>
-                    Flaxseed contains omega-3 fatty acids. Research suggests that these
-                    may help prevent different types of cancer cells from growing.
-                    Flaxseed also contains lignans, which are antioxidants that may slow
-                    tumor growth by preventing them from forming new blood vessels.
-                </p>
-                <p>
-                    One <a href="https://link.springer.com/article/10.1007/s10552-013-0155-7">
-                        2013</a> survey found a lower incidence of breast cancer among
-                    females who consumed flaxseed regularly.
-                </p>
-                <p>
-                    Also, in 2018, the
-                    authors of a <a href="https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5808339/">
-                        review</a> concluded that flaxseed may help
-                    reduce the risk of breast cancer after menopause.
-                </p>
-                <p>
-                    Lignans are a type of phytoestrogen, which is a plant-based nutrient that
-                    acts in a similar way to estrogen. There has been some concern that
-                    phytoestrogens may increase the risk of breast cancer, but recent research
-                    suggests that they may play a protective role.
-                </p>
-                <h3>Improving cholesterol and heart health:</h3>
-                <p>
-                    The American Heart Association (AHA) recommend eating more fiber and omega-3s
-                    to boost heart health. Lignans, too, may help prevent cardiovascular disease.
-                    Flaxseed contains all of these nutrients.
-                </p>
-                <p>
-                    Flaxseed also contains phytosterols. Phytosterols have a similar structure
-                    to cholesterol, but they help prevent the absorption of cholesterol in the
-                    intestines.
-                </p>
-                <p>
-                    Consuming phytosterols may therefore help reduce levels of low-density
-                    lipoprotein (LDL), or “bad,” cholesterol in the body.
-                    In 2010, researchers looked at the effect of flaxseed on the cholesterol
-                    levels of males with moderately high cholesterol. Participants took either
-                    a 20 milligram (mg) capsule containing lignans, a 100 mg capsule, or a
-                    placebo for 12 weeks.
-                </p>
-                <p>
-                    Cholesterol levels fell after taking lignans, especially in those who took
-                    the 100 mg capsules.
-                </p>
-                <p>
-                    The researchers behind a 2012 study involving 17 people found that consuming
-                    flaxseed lowered LDL cholesterol levels and helped the body remove fat,
-                    although they note that the overall diet may also play a role. The team
-                    suggested that dietary flaxseed may be useful for lowering cholesterol
-                    levels.
-                </p>
-                <p>
-                    Some scientists have also linked omega-3 oils, which are usually present
-                    in oily fish, to reductions in cardiovascular risk. Researchers have
-                    suggested that flaxseed could offer an alternative to marine sources
-                    of omega 3. This could make it a useful resource for people who follow
-                    a plant-based diet.
-                </p>
-                <h3>Easing the symptoms of Arthiritis:</h3>
-                <p>
-                    According to the Arthritis Foundation, flaxseed may help reduce joint
-                    pain and stiffness. Some people take it for rheumatoid arthritis, lupus,
-                    and Raynaud's phenomenon.
-
-                    They add that there is a lack of evidence to support its use for this
-                    purpose, but they say that the ALA in flaxseed may help reduce
-                    inflammation. People can take it:
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>Ground (one tablespoon per day)</li>
-                        <li>As an oil (one to three tablespoons per day)</li>
-                        <li>In capsules (1300-3000 mg per day)</li>
-                    </ul>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_SideEffects() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>What are the side effects of using Flax Seeds?</h2>
-                    <p>
-                        Because of the high fiber content, flaxseeds may result in increased
-                        bowel movements. They may also cause digestive issues, such as:
-                    </p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>Bloating</li>
-                            <li>Cramps</li>
-                            <li>Gas</li>
-                            <li>Abdominal pain</li>
-                        </ul>
-                        <ul>
-                            <li>Constipation</li>
-                            <li>Diarrhea</li>
-                            <li>Stomachache</li>
-                            <li>Nausea</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    You can avoid these side effects by moderately consuming flaxseeds.
-                    It is most often an excessive intake that causes these digestive
-                    issues. Also, ensure that you drink plenty of water throughout the
-                    day when you have eaten flaxseeds. This will keep problems, such as
-                    constipation, away.
-                </p>
-                <p>
-                    Raw and unripe flaxseeds may contain toxic compounds. Hence, avoid
-                    using them. Pregnant women should avoid flaxseeds, even flaxseed
-                    oil or their supplements, due to the mild estrogenic (hormonal) effects.
-                    It is rare for someone to develop allergic reactions to flaxseeds.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Dosage() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>What is the recommended dosage for using Flax Seeds?</h2>
-                    <p>
-                        Consume 3 to 4 tablespoons of flaxseed daily to benefit from the
-                        fiber and omega-3 fatty acids. If you are taking flaxseed for
-                        constipation, consume 2 tablespoons daily. Toddlers should consume
-                        about 1 tablespoon per day. While there are no reports of flaxseed
-                        overdosing, you may experience excess gas and bloating if you eat
-                        large servings of the meal because of its fiber and fat content.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function GrowIngredient() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Flax Seeds"
-                    familyName="Linaceae" origin="Syria, Turkey"
-                    geoLocation="Syria, Switzerland, China, and India" />
-                <div className="content-wrapper">
-                    <h2>How do I grow Flax Seeds?</h2>
-                    <p>
-                        Growing Flax Seeds on a commercial level
-                        can be a complicated process, but planting flax from seed in your
-                        garden is easier than you may think. In fact, you have likely grown
-                        its wildflower cousins, blue flax and scarlet flax before, or know
-                        someone who has. Common flax, like its cousins, is a cool-season
-                        plant, and the seeds should be planted as soon as the ground can
-                        be worked in spring.
-                    </p>
-                    <p>
-                        A late frost will not usually harm the plants
-                        once they emerge, as seedlings with at least two leaves can tolerate
-                        temperatures as low as 28 F. (-2 C.). Look for a sunny, sheltered
-                        planting site when planting flax from seed. Although flax will
-                        adapt to most well-drained soil types, rich soil is optimum.
-                        Dig in a generous amount of compost, manure, or other organic
-                        matter, especially if your soil is poor.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function FlaxSeeds() {
-    return (
-        <>
-            <IngredientNameHeading ingredientName="Flax Seeds" />
-            <div className="ingredient-hub">
-                <ul>
-                    <li>Overview</li>
-                    <li>How to use</li>
-                    <li>Benefits</li>
-                    <li>Side Affects</li>
-                    <li>Dosage</li>
-                    <li>Grow Flax Seeds</li>
-                </ul>
-            </div>
-            <div className="component-content-rendering">
-                {/* Components will be rendered here using react-router-dom <Link> components*/}
-                <Ingredient_Overview />
-                <Ingredient_HowToUse />
-                <Ingredient_Benefits />
-                <Ingredient_SideEffects />
-                <Ingredient_Dosage />
-                <GrowIngredient />
-            </div>
-        </>
-    )
-}
-
-export default FlaxSeeds;
