+++ src/components/layout/Header/Ingredients-content/Plants/FlaxSeeds/FlaxSeeds.tsx
@@ -1,0 +1,38 @@
+import { Link, Outlet } from "react-router-dom";
+import "../../../../../../App.css";
+import IngredientNameHeading from "../../Ingredient-heading";
+
+function FlaxSeeds() {
+    return (
+        <>
+            <IngredientNameHeading ingredientName="Flax Seeds" />
+            <div className="ingredient-hub">
+                <ul>
+                    <li>
+                        <Link to="/FlaxSeeds_Overview">Overview</Link>
+                    </li>
+                    <li>
+                        <Link to="/FlaxSeeds_HowToUse">How to use</Link>
+                    </li>
+                    <li>
+                        <Link to="/FlaxSeeds_Benefits">Benefits</Link>
+                    </li>
+                    <li>
+                        <Link to="/FlaxSeeds_SideEffects">Side Effects</Link>
+                    </li>
+                    <li>
+                        <Link to="/FlaxSeeds_Dosage">Dosage</Link>
+                    </li>
+                    <li>
+                        <Link to="/Grow_FlaxSeeds">Grow Flax Seeds</Link>
+                    </li>
+                </ul>
+            </div>
+            <div className="component-content-rendering">
+                <Outlet />
+            </div>
+        </>
+    )
+}
+
+export default FlaxSeeds;
