--- src/components/layout/Header/Ingredients-content/Plants/Ginseng.tsx
@@ -1,427 +1,0 @@
-import Ingredient from "../Ingredient";
-import IngredientNameHeading from "../Ingredient-heading";
-
-function Ingredient_Overview() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="Asia, Japan and Canada" />
-                <div className="content-wrapper">
-                    <h2>What is Ginseng?</h2>
-                    <p>
-                        Ginseng is a name you have probably heard being thrown around,
-                        especially in relation to natural healthcare products and supplements.
-                        But there are actually many different species of ginseng, found all
-                        over the world. There is Panax ginseng(Korean
-                        ginseng), which is present in Asia. Then Panax quinquefolius(American
-                        ginseng) as it is found naturally in Canada and the  USA.
-                        Another is Panax japonicus(Japanese ginseng).
-                        Panax ginseng originated in Asia, and today it is grown in the cold
-                        temperate regions in North East China and Korean peninsula. It was
-                        first recorded in Chinese medicine over 2000 years ago, and is
-                        now one of the most heavily cultivated herbs in the world.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    Panax ginseng is a perennial shrub, about 50-80cm in height. It
-                    bears inconspicuous flowers which develop into red berries, and has
-                    a turnip-shaped root that quite often, due to the way it branches,
-                    appears like a human figure. Panax Ginseng  is a species of plant
-                    whose root is the original source of ginseng. It is a perennial
-                    plant that grows in the mountains of East Asia.
-                </p>
-                <p>
-                    It is the root that is of the most interest medicinally. That is
-                    because most of the important components and active ingredients are
-                    found here.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_HowToUse() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="Asia, Japan and Canada" />
-                <div className="content-wrapper">
-                    <h2>How is Ginseng used?</h2>
-                    <p>Ginseng comes in the form of:</p>
-                    <div>
-                        <ul>
-                            <li>Roots</li>
-                            <li>Extracts</li>
-                            <li>Powders</li>
-                            <li>Tablets</li>
-                            <li>Capsules</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>Roots, extracts and powders:</h3>
-                <p>
-                    Ginseng root, extract, and  powder can be dissolved in water, juice, or
-                    smoothies. Powders are often in higher dosages than tablets or capsules.
-                    You may also find smaller amounts of ginseng added to energy drinks
-                    and herbal teas.
-                </p>
-                <p>
-                    While not as common, you can consume ginseng in its plant form by:
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>Peeling and chewing the raw root</li>
-                        <li>Soaking the peeled raw root in wine to create an extract</li>
-                        <li>Boiling the peeled, raw root to make tea</li>
-                        <li>Soaking or boiling dried ginseng root to create an extract.</li>
-                    </ul>
-                </div>
-                <h3>Tablets and Capsules:</h3>
-                <p>
-                    Tablets or capsules typically contain the ground-up root or an extract
-                    of one or more types of ginseng root. People are usually advised to
-                    take one to two ginseng capsules twice daily with food. The number
-                    of capsules may vary depending on the supplement dosage.
-                </p>
-            </div>
-        </>
-    )
-}
-function Ingredient_Benefits() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="Asia, Japan and Canada" />
-                <div className="content-wrapper">
-                    <h2>What are the benefits of using Ginseng?</h2>
-                    <p>
-                        Ginseng is an herb that is rich in antioxidants. It may offer benefits
-                        for brain health, immune function, blood sugar control, and more.
-                        Ginseng has also been used in traditional Chinese medicine for centuries.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h2>Health benefits include:</h2>
-                <h3>Contains antioxidants that reduces inflammation:</h3>
-                <p>
-                    Ginseng has beneficial antioxidant and anti-inflammatory properties.
-                    Some test-tube studies have shown that ginseng extracts and ginsenoside c
-                    ompounds could inhibit inflammation and reduce oxidative damage to cells,
-                    which can contribute to chronic disease.
-                </p>
-                <p>
-                    The results are promising in humans, as well.
-                    One study in 12 active males found that short-term supplementation with
-                    American ginseng extract reduced exercise-induced muscle damage and decreased
-                    markers of inflammation compared to a placebo.
-                </p>
-                <p>
-                    A larger study in 2014 followed 71 postmenopausal females who took 3 grams(g)
-                    of red ginseng or a placebo daily for 12 weeks. Antioxidant activity and
-                    oxidative stress markers were then measured. red ginseng may help reduce
-                    oxidative stress by increasing antioxidant enzyme activities.
-                </p>
-                <h3>May benefit brain function:</h3>
-                <p>
-                    Ginseng could help improve brain functions like memory, behavior, and mood.
-                    Some test-tube and animal studies show that components in ginseng, like
-                    ginsenosides and compound K, could protect the brain against damage caused
-                    by free radicals.
-
-                    Interestingly, one study in 6,422 older adults found that regular
-                    consumption of ginseng for at least 5 years was associated with improved
-                    cognitive function later in life.
-
-                    Another small study showed that taking 200 milligrams (mg) of American
-                    ginseng significantly improved working memory after 3 hours, compared to a
-                    placebo.
-
-                    According to one review, ginseng could also help ease stress and may
-                    offer benefits for depression and anxiety.
-                </p>
-                <h3>Could improve erectile dysfunction:</h3>
-                <p>
-                    Some research suggests that ginseng may be a useful alternative for the
-                    treatment of erectile dysfunction (ED).
-
-                    According to some older research, certain compounds found in ginseng may
-                    protect against oxidative stress in the blood vessels and tissues of the
-                    penis to help restore normal function.
-
-                    Additionally, studies have shown that ginseng may promote the production
-                    of nitric oxide, a compound that improves muscle relaxation in the male
-                    genitals and increases blood circulation.
-
-                    Still, research on the effects of ginseng on ED has turned up mixed
-                    results and more high quality studies are needed.
-
-                    In fact, one review of nine studies found that ginseng could improve
-                    self-reported ability to have intercourse in males.
-                </p>
-                <h3>May boost the immune system:</h3>
-                <p>
-                    According to one review, ginseng possesses potent anti-bacterial,
-                    anti-fungal, and anti-viral properties and may enhance the function of
-                    the immune system.
-                </p>
-                <p>
-                    Interestingly, one animal study found that black red ginseng extract
-                    increased the number of immune cells and enhanced antioxidant levels
-                    in the liver.
-                </p>
-                <p>
-                    Similarly, another study in 100 people showed that taking 2 g of Korean
-                    red ginseng per day for 8 weeks significantly increased levels of
-                    immune cells compared to a placebo.
-                </p>
-                <h3>May have potential benefits against cancer:</h3>
-                <p>
-                    Ginseng may be helpful in reducing the risk of certain cancers.
-                    Ginsenosides in this herb have been shown to help reduce inflammation and
-                    provide antioxidant protection.
-
-                    The cell cycle is the process by which cells normally grow and divide.
-                    Ginsenosides could benefit this cycle by preventing abnormal cell
-                    production and growth.
-
-                    A review of several studies concluded that people who take ginseng may
-                    have a a 16% lower risk of developing cancer. Ginseng may also help
-                    improve the health of people undergoing chemotherapy and could reduce
-                    side effects and enhance the effect of some cancer treatments
-                </p>
-                <h3>May fight tiredness and increase energy levels:</h3>
-                <p>
-                    Ginseng has been shown to help ease fatigue and increase energy levels.
-
-                    Various animal studies have linked some components in ginseng, like
-                    polysaccharides and oligopeptides, with lower oxidative stress and higher
-                    energy production in cells, which could help decrease fatigue.
-
-                    One review of 10 studies concluded that ginseng could significantly
-                    improve symptoms of chronic fatigue syndrome compared to a placebo,
-                    even after just 15 days.
-
-                    Another review showed that taking American or Asian ginseng could
-                    decrease symptoms of cancer-related fatigue when taken in doses of
-                    2,000 mg or 3,000 mg per day, respectively.
-                </p>
-                <h3>Could lower blood sugar:</h3>
-                <p>
-                    Ginseng seems to be beneficial in the control of blood sugar levels
-                    in people both with and without diabetes.
-                </p>
-                <p>
-                    American and Asian ginseng have been shown to improve pancreatic
-                    cell function, boost insulin production, and enhance the uptake of
-                    blood sugar in tissues.
-                </p>
-                <p>
-                    Moreover, studies show that ginseng extracts help by providing
-                    antioxidant protection that can help reduce free radicals in the
-                    cells of those with diabetes.
-                </p>
-                <p>
-                    One review of eight studies found that ginseng supplementation
-                    could decrease fasting blood sugar levels and improve insulin
-                    sensitivity in people with type 2 diabetes
-                </p>
-                <p>
-                    Another 8-week study showed that taking 3 g per day of American ginseng
-                    significantly reduced fasting blood sugar levels and improved hemoglobin
-                    A1c, a marker of long-term blood sugar control, compared to a placebo in
-                    people with type 2 diabetes.
-                </p>
-                <p>
-                    It seems that fermented red ginseng could be even more effective at
-                    blood sugar control. Fermented ginseng is produced with the help of live
-                    bacteria that transform the ginsenosides into a more easily absorbed and
-                    potent form.
-                </p>
-            </div>
-        </>
-    )
-}
-function Ingredient_SideEffects() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="Asia, Japan and Canada" />
-                <div className="content-wrapper">
-                    <h2>What are the side effects of using Ginseng?</h2>
-                    <p>
-                        People who take more than 3000 mg of
-                        American ginseng and 3g of Asian ginseng per day are more likely to
-                        encounter side effects. Taking ginseng for prolonged periods may also
-                        increase your risk of side effects.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    Sleep problems are the most common side effect of taking Asian and
-                    American ginseng. Other side effects of ingesting ginseng include:
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>Headache</li>
-                        <li>Nausea</li>
-                        <li>Diarrhea</li>
-                        <li>Changes in blood pressure</li>
-                    </ul>
-                    <ul>
-                        <li>Increased heart rate</li>
-                        <li>Loss of appetite</li>
-                        <li>Breast pain</li>
-                        <li>Menstrual problems</li>
-                    </ul>
-                </div>
-            </div>
-        </>
-    )
-}
-function Ingredient_Dosage() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="Asia, Japan and Canada" />
-                <div className="content-wrapper">
-                    <h2>What is the recommended dosage of using Ginseng?</h2>
-                    <p>
-                        There is no standard ginseng dosage recommendation. The dosage depends
-                        on the type of ginseng and the amount of ginsenosides. Research suggests
-                        adults may safely take 100-3000 mg of American ginseng. Safe Asian ginseng
-                        dosages range from 200mg to 3g per day.
-                    </p>
-                    <h3>Can you consume too much Ginseng?</h3>
-                    <p>
-                        Yes, you can take too much ginseng and experience uncomfortable side
-                        effects. However, there has not been enough testing to prove a toxic
-                        amount of ginseng for adults.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-function GrowIngredient() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginseng"
-                    familyName="Araliaceae" origin="Shangdang China"
-                    geoLocation="China, Korea, Japan, and Canada" />
-                <div className="content-wrapper">
-                    <h2>How do I grow Ginseng?</h2>
-                    <p>First of all there are three types of Ginseng:</p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>Asian Ginseng</li>
-                            <li>Japanese Ginseng</li>
-                            <li>American Ginseng</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>How to plant Ginseng:</h3>
-                <p>
-                    To prepare your site, remove low branches of trees to promote air movement
-                    and be sure your soil has plenty of incorporated organic matter. Well-composted
-                    leaf litter is an excellent addition.
-                </p>
-                <p>
-                    Ginseng can be grown from seed or seedling. Purchase good quality seed that
-                    has been stratified (gone through a cold period) for fastest germination,
-                </p>
-                <p>
-                    There is an advantage to starting with ginseng seedlings instead of seeds
-                    which are more widely available these days and may be planted in spring or
-                    fall. One or two year old seedlings will speed up the time to maturity, when
-                    the root is large enough to harvest, which is a minimum of five years old.
-                </p>
-                <h3>Can I grow Ginseng in a container?</h3>
-                <p>
-                    Ginseng may be grown in a container, especially in its early years and as
-                    seedlings. Select a container that has a drainage hole. Use a loamy soil
-                    mix and plant your seeds about one-and-a-half inches deep. Site the
-                    container in a shady spot, keep it watered, and leave it outdoors in
-                    winter. Bury [the container] a few inches into the ground and mulch to
-                    mimic a wild plant going through a cold period mulched with litter from the
-                    fall of autumn leaves.
-                </p>
-                <p>
-                    Ginseng is not a good choice for growing indoors. It prefers a well-drained
-                    forest soil, good air flow, and needs to go through the cycles of the seasons
-                    including cold dormancies, where the leaves and stem of the plant die back
-                    to the ground in the fall, to reemerge in the spring.
-                </p>
-                <h3>Harvesting Ginseng:</h3>
-                <p>
-                    Ginseng is a slow grower, even in the best of conditions. Ginseng plants
-                    may grow for decades and can be aged by counting the neck scars left on the
-                    top of the root from each growing season's stem.
-                </p>
-                <p>
-                    It will likely be 5-10 years before you can harvest a decent root. Roots should
-                    be dug in the fall as the leaves begin to die. Gently wash off your harvested
-                    roots, then place them on a screen to dry in a well-ventilated area out of
-                    direct sunlight. Rotate them occasionally so they dry evenly. Drying usually
-                    takes two to four weeks, although large roots may take longer. Store your roots
-                    in a ventilated box or basket.
-                </p>
-                <h3>Why is there a red and white Ginseng?</h3>
-                <p>
-                    The difference in colour comes down to how ginseng is treated. White ginseng
-                    is naturally dried in the sun. Red ginseng is steam heated between 100-110
-                    degrees for 2-3 hours. This treatment kills any potential pathogens on the
-                    plant's roots. The downside is that it degrades some of the active components
-                    which are found in the outer layers of bark.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ginseng() {
-    return (
-        <>
-            <IngredientNameHeading ingredientName="Ginseng" />
-            <div className="ingredient-hub">
-                <ul>
-                    <li>Overview</li>
-                    <li>How to use</li>
-                    <li>Benefits</li>
-                    <li>Side Affects</li>
-                    <li>Dosage</li>
-                    <li>Grow Ginseng</li>
-                </ul>
-            </div>
-            <div className="component-content-rendering">
-                {/* Components will be rendered here using react-router-dom <Link> components*/}
-                <Ingredient_Overview />
-                <Ingredient_HowToUse />
-                <Ingredient_Benefits />
-                <Ingredient_SideEffects />
-                <Ingredient_Dosage />
-                <GrowIngredient />
-            </div>
-        </>
-    )
-}
-
-export default Ginseng;
