--- src/components/layout/Header/Ingredients-content/Plants/Licorice.tsx
@@ -1,393 +1,0 @@
-import Ingredient from "../Ingredient";
-import IngredientNameHeading from "../Ingredient-heading";
-import "../../../../../App.css";
-function Ingredient_Overview() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>What is Licorice?</h2>
-                    <p>
-                        The liquorice plant is an herbaceous perennial legume native to Western
-                        Asia, North Africa, and Southern Europe. Liquorice is used as a
-                        flavouring in candies, tobacco, beverages, and pharmaceuticals,
-                        and is marketed as a dietary supplement.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-function Ingredient_HowToUse() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>How is Licorice used?</h2>
-                    <p>
-                        Licorice is used:
-                    </p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>As a flavoring in candy</li>
-                            <li>As an ingredient for tabacco products</li>
-                            <li>In medicine</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>Flavoring in candy:</h3>
-                <h4>
-                    <a target="_blank"
-                        href="https://licorice.com/blogs/news/learn-about-the-different-types-of-licorice#:~:text=Licorice%20comes%20in%20so%20many,%2C%20salty%2C%20and%20sour%20licorice.">Licorice
-                        with its favorite candies:</a>
-                </h4>
-                <ul>
-                    <li>Red Licorice</li>
-                    <li>Black Licorice</li>
-                    <li>Salty Licorice</li>
-                    <li>Sour Licorice</li>
-                </ul>
-                <h3>Tabacco products:</h3>
-                <p>
-                    Licorice is known for its sweet and slightly anise-like flavor. In the
-                    tobacco industry, it may be used as a flavoring agent to impart a
-                    distinctive taste to tobacco products. Licorice flavoring can be found in
-                    various types of tobacco products, including cigarettes, cigars, and pipe
-                    tobacco.
-                </p>
-                <p>
-                    Licorice extract, often referred to as licorice juice or
-                    licorice molasses, contains glycyrrhizinic acid, which can help retain moisture
-                    in tobacco. This moisture-retaining property can affect the texture and burn
-                    characteristics of tobacco products.
-                </p>
-                <p>
-                    Licorice may be blended with other herbs, spices, or
-                    flavorings to create unique and aromatic tobacco blends. These blends are
-                    often used in herbal smoking products and are marketed as alternatives to
-                    traditional tobacco.
-                </p>
-                <h3>Medicine:</h3>
-                <p>
-                    What is Licorice Root? Well it is the root of the plant which is often
-                    added to herbal teas to enhance the flavor which has been used for centuries
-                    in ancient like Egypt, Europe, and Asia. You can find licorice root
-                    tea in health food stores and some supermarkets. You may be able to
-                    find licorice root in some health food shops if you would like to make
-                    the tea yourself.
-                </p>
-            </div >
-        </>
-    )
-}
-function Ingredient_Benefits() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>What are the benefits of using Licorice?</h2>
-                    <p>
-                        Licorice root is the root of the plant which is often
-                        added to herbal teas to enhance the flavor which has been
-                        used for centuries in ancient like Egypt, Europe, and Asia.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h2>Health benefits:</h2>
-                <h3>Licorice Tea</h3>
-                <p>
-                    There is a long tradition but not a lot of research.
-                    There is not enough high-quality evidence to clearly support its use for
-                    any health condition. The compounds in licorice root that may be helpful
-                    include glycyrrhizin (which is 50 times sweeter than sugar), licoricidin,
-                    and liquiritin.
-                </p>
-                <p>
-                    Sore throat: For centuries, people have sipped licorice root tea to help
-                    soothe a sore throat. A licorice compound called licoricidin has been
-                    shown to have antibacterial properties that may help.
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>
-                            <h3>It Could Help Your Teeth:</h3>
-                            <dt>
-                                Antibacterial properties in licorice root extract might have
-                                potential to help prevent cavities. That is not certain,
-                                though.
-                            </dt>
-                            <dt>
-                                Researchers did a preliminary study using sugar-free
-                                lollipops containing the extract to see if it was worth studying
-                                further. It did help the preschoolers who used the lollipops twice
-                                daily.
-                            </dt>
-                            <dt>
-                                However, more studies are needed to see how well this strategy works.
-                                Two licorice compounds, licoricidin and glabridin, may be what helps
-                                counter cavities.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>It May Help With Digestion:</h3>
-                            <dt>
-                                One small study focused exclusively on a product made with a compound
-                                extracted from licorice root. The researchers studied it in 50 people
-                                with functional dyspepsia, which includes symptoms like heartburn,
-                                nausea, and upset stomach.
-                            </dt>
-                            <dt>
-                                Half of the group took the product, the other half got a placebo,
-                                and they did not know which they would gotten. Those who took the
-                                product got more relief from their symptoms than those who got the
-                                placebo. The study did not show why that was.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>It May Help With Eczema:</h3>
-                            <dt>
-                                Some studies have shown that gels that you put on your skin made
-                                with licorice root can help treat atopic dermatitis (also known
-                                as eczema). Licorice root can help reduce irritated, inflamed,
-                                and itchy skin caused by
-                                <a href="https://www.webmd.com/skin-problems-and-treatments/eczema/atopic-dermatitis-eczema">
-                                    eczema</a>.
-                            </dt>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </>
-    )
-}
-function Ingredient_SideEffects() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>What are the side effects of using Licorice?</h2>
-                    <p>
-                        If you know you have genuine licorice, it is best to eat it
-                        in small amounts. Too much licorice, or using it for too long,
-                        can raise your blood pressure. It can also drop your potassium
-                        levels, and that could cause an abnormal heart rhythm.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-function Ingredient_Dosage() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>What is the recommended dosage for using Licorice?</h2>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    If you are pregnant, having a lot of licorice is linked to a greater
-                    chance of premature birth and health problems in the baby.
-                </p>
-                <p>
-                    The risk may stem from the glycyrrhizin in licorice. Too much of it
-                    can cause serious health problems. Exactly how much is too much isn
-                    not clear because the amount of licorice in a product can vary.
-                    The risks with potassium and blood pressure are greatest for people
-                    who already have high blood pressure, heart disease, or kidney disease.
-                </p>
-                <p>
-                    For people who are 40 or older, eating at least 2 ounces of genuine
-                    black licorice daily for 2 weeks might lead to hospitalization for
-                    heart palpitations. Genuine licorice can also interfere with some
-                    medicines and supplements, so be sure to talk with your doctor if
-                    you notice any changes in how your medicines or supplements work.
-                </p>
-            </div>
-        </>
-    )
-}
-function GrowIngredient() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Liquorice(Licorice)"
-                    familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
-                    and Southern Europe"
-                    geoLocation="Western Asia, North Africa, 
-                    and Southern Europe" />
-                <div className="content-wrapper">
-                    <h2>How do I grow Licorice?</h2>
-                    <h3>Where to Plant Licorice Plant:</h3>
-                    <p>
-                        While it's most commonly grown as an annual, licorice plant is a
-                        woody tropical perennial that is winter hardy in Zones 9-11.
-                        In the rest of the country, either plant it as an annual or in a
-                        container that can be moved inside in cold temperatures.
-                        Licorice plant is useful at the front of sunny beds and borders
-                        and can be planted on slopes for erosion control. It also is a
-                        good choice for containers and hanging baskets.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>How and When to Plant Licorice Plant:</h3>
-                <p>
-                    Plant licorice plant seeds (if you can find them) in late autumn in a
-                    greenhouse or the house to develop plants that can go into the garden
-                    when the weather warms the following spring. Gardeners who prefer to
-                    buy nursery plants should put them out in the garden at the same time
-                    they put out tomato plants. If you have a potted licorice plant, you
-                    can take stem cuttings and root them in water timed for a late spring
-                    planting outside.
-                </p>
-                <div className="ingredient-benefits">
-                    <ol>
-                        <li>In the garden, dig a hole as deep as your plant's nursery pot.</li>
-                        <li>Remove the plant from the pot and loosen the soil around the roots. </li>
-                        <li>Place the plant in the hole and backfill with soil.</li>
-                        <li>Space multiple plants about 30 inches apart.</li>
-                        <li>
-                            Water new plants well to settle the soil around roots and help
-                            the plants get established.
-                        </li>
-                    </ol>
-                </div>
-                <h3>Light:</h3>
-                <p>
-                    Licorice plant is a fast grower and likes as much sun as possible.
-                    In part shade, plants can become leggy and need pruning to keep from
-                    looking messy. Also, shade-grown plants don't look quite as silvery, as the
-                    hairs are not as dense when the plant grows in the shade.
-                </p>
-                <h3>Soil and Water:</h3>
-                <p>
-                    When looking for a home for your licorice plant, make sure to plant it
-                    in well-drained soil. Licorice plant doesn't appreciate sitting in too
-                    much water. If it does, the plant may begin to rot. Once the plant is
-                    established, it is drought-tolerant, although it prefers regular watering.
-                </p>
-                <h3>Temperature and Humidity:</h3>
-                <p>
-                    Licorice plants grow best in dry, hot environments that receive little rainfall.
-                    They do not tolerate any frost. Bring in the plants before the first frost
-                    to winter them over, or take stem cuttings to begin a new crop for the
-                    following spring.
-                </p>
-                <h3>Fertilizer:</h3>
-                <p>
-                    Licorice plants are tolerant of many soil conditions and don't require
-                    much fertilizer, but if the soil is poor, add compost or organic material
-                    to enrich it. Then, a mid-season application of a balanced fertilizer
-                    is all the plants need. For the amount to use, follow product label
-                    instructions.
-                </p>
-                <h3>Pruning:</h3>
-                <p>
-                    Licorice plant handles pruning well. It's a good idea to give trailing
-                    varieties a pinch early in their growth to encourage good branching.
-                </p>
-                <h3>Potting and Repotting Licorice Plant:</h3>
-                <p>
-                    This plant's semi-trailing or cascading growth habit works well in
-                    containers and hanging baskets. Some varieties have a more upright habit,
-                    so if you intend to use the plant as a spiller, check its habit before
-                    purchasing. Repot a licorice plant yearly in a somewhat large container
-                    being careful not to disturb the roots.
-                </p>
-                <h3>Pests and Problems:</h3>
-                <p>
-                    Licorice plant is relatively pest-free, but watch for aphids, spider
-                    mites and whiteflies, all of which can be treated with neem oil.
-                    This plant likes hot, dry weather. If the plant is grown in a wet
-                    area, it may develop root rot. If water stands on its leaves or the
-                    plant grows in an area with high humidity, botrytis can occur.
-                </p>
-                <h3>How to Propagate Licorice Plant:</h3>
-                <p>
-                    Licorice plant can be propagated by seeds, stem cuttings and root divisions.
-                </p>
-                <p>
-                    In early spring in all but the warmest zones, prepare seeds by soaking them in
-                    water for 24 hours. Then, sow the seeds on top of a seed-starting mix and do
-                    not cover them. Supply bottom heat at about 68 degrees Fahrenheit, and
-                    seedlings should appear two to three weeks later. Wait until the weather
-                    warms to set them out in the garden. In Zones 9-11, sow seeds outside
-                    in early spring in a full sun site. Cover them lightly with sand to keep
-                    them from blowing away or being eaten by birds.
-                </p>
-                <p>
-                    If you have a licorice plant and want more just like it, take stem cuttings
-                    or root divisions for an exact duplicate of this hybrid plant. Take 6-inch
-                    stem cuttings and remove the leaves from the bottom half. Dip it into rooting
-                    hormone and settle it in a small pot of moist potting soil. Mist it every
-                    day and tug on it slightly in a couple of weeks to determine whether it is
-                    rooted.
-                </p>
-                <p>
-                    The hands-down easiest way to make more licorice plants is to divide one.
-                    Use a sharp shovel and cut a mature plant into four sections, each containing
-                    a portion of leaves and roots. Replant the divisions immediately.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Licorice() {
-    return (
-        <>
-            <IngredientNameHeading ingredientName="Liquorice(Licorice)" />
-            <div className="ingredient-hub">
-                <ul>
-                    <li>Overview</li>
-                    <li>How to use</li>
-                    <li>Benefits</li>
-                    <li>Side Affects</li>
-                    <li>Dosage</li>
-                    <li>Grow Licorice</li>
-                </ul>
-            </div>
-            <div className="component-content-rendering">
-                {/* Components will be rendered here using react-router-dom <Link> components*/}
-                <Ingredient_Overview />
-                <Ingredient_HowToUse />
-                <Ingredient_Benefits />
-                <Ingredient_SideEffects />
-                <Ingredient_Dosage />
-                <GrowIngredient />
-            </div>
-        </>
-    )
-}
-
-export default Licorice;
