+++ src/components/layout/Header/Ingredients-content/Plants/Licorice/Licorice_SideEffects.tsx
@@ -1,0 +1,55 @@
+import { Link, Outlet } from "react-router-dom";
+import "../../../../../../App.css";
+import Ingredient from "../../Ingredient";
+import IngredientNameHeading from "../../Ingredient-heading";
+
+function Licorice_SideEffects() {
+    return (
+        <>
+            <IngredientNameHeading ingredientName="Licorice" />
+            <div className="ingredient-hub">
+                <ul>
+                    <li>
+                        <Link to="/Licorice_Overview">Overview</Link>
+                    </li>
+                    <li>
+                        <Link to="/Licorice_HowToUse">How to use</Link>
+                    </li>
+                    <li>
+                        <Link to="/Licorice_Benefits">Benefits</Link>
+                    </li>
+                    <li>
+                        <Link to="/Licorice_SideEffects">Side Effects</Link>
+                    </li>
+                    <li>
+                        <Link to="/Licorice_Dosage">Dosage</Link>
+                    </li>
+                    <li>
+                        <Link to="/Grow_Licorice">Grow Licorice</Link>
+                    </li>
+                </ul>
+            </div>
+            <div className="component-content-rendering">
+                <Outlet />
+                <div className="flex-content">
+                    <Ingredient ingredientName="Liquorice(Licorice)"
+                        familyName="Fabaceae" origin="Native to Western Asia, North Africa, 
+                    and Southern Europe"
+                        geoLocation="Western Asia, North Africa, 
+                    and Southern Europe" />
+                    <div className="content-wrapper">
+                        <h2>What are the side effects of using Licorice?</h2>
+                        <p>
+                            If you know you have genuine licorice, it is best to eat it
+                            in small amounts. Too much licorice, or using it for too long,
+                            can raise your blood pressure. It can also drop your potassium
+                            levels, and that could cause an abnormal heart rhythm.
+                        </p>
+                    </div>
+                </div>
+            </div>
+        </>
+    )
+}
+
+export default Licorice_SideEffects;
