--- src/components/layout/Header/Ingredients-content/Spices/Ginger.tsx
@@ -1,336 +1,0 @@
-import "../../../../../App.css"
-import Ingredient from "../Ingredient";
-import IngredientNameHeading from "../Ingredient-heading";
-
-function Ingredient_Overview() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>What is Ginger?</h2>
-                    <p>
-                        Ginger root or ginger, is widely used as a spice and a folk medicine.
-                        It is a herbaceous perennial which grows annual pseudostems (false stems
-                        made of the rolled bases of leaves) about one meter tall, bearing narrow
-                        leaf blades. The inflorescences bear flowers having pale yellow petals
-                        with purple edges, and arise directly from the rhizome on separate shoots.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    Ginger is in the family Zingiberaceae, which also includes turmeric (Curcuma
-                    longa), cardamom (Elettaria cardamomum), and galangal. Ginger originated
-                    in Southeast Asia and was likely domesticated first by the
-                    Austronesian peoples. It was transported with them throughout the Indo-Pacific
-                    during the Austronesian expansion (c.5,000 BP), reaching as far as Hawaii.
-                </p>
-                <p>
-                    Ginger is one of the first spices to have been exported from Asia, arriving
-                    in Europe with the spice trade, and was used by ancient Greeks and Romans.
-                    The distantly related dicots in the genus Asarum are commonly called wild
-                    ginger because of their similar taste.
-                </p>
-                <p>
-                    Used in traditional medicine in China, India and Japan for centuries, and
-                    as a dietary supplement, research shows that ginger may be helpful for
-                    mild nausea and vomiting associated with pregnancy, although its safety
-                    has not been demonstrated. It remains uncertain whether ginger is effective
-                    for treating any disease, and use of ginger as a drug has not been approved
-                    by the FDA.
-                </p>
-                <p>
-                    In 2020, world production of ginger was 4.3 million tonnes, led by India
-                    with 43% of the world total.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_HowToUse() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>How is Ginger used?</h2>
-                    <p>
-                        Ginger has been used for thousands of years, even now
-                        where it was used for:
-                    </p>
-                    <div>
-                        <ul>
-                            <li>Foods, Beverages, and Desserts</li>
-                            <li>Beauty, and Cosmetics</li>
-                            <li>Food decoration</li>
-                            <li>Aromatherapy</li>
-                            <li>Medicine which is the mostpopular use in the
-                                far east such as Asia.
-                            </li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Benefits() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>What are the benefits of using Ginger?</h2>
-                    <p>
-                        Ginger, the "root" or the rhizome, of the plant Zingiber officinale,
-                        has been a popular spice and herbal medicine for thousands of years.
-                        It has a long history of use in Asian, Indian, and Arabic herbal traditions.
-                    </p>
-                    <p>
-                        In China, for example, ginger has been used to help digestion and treat
-                        stomach upset, diarrhea, and nausea for more than 2000 years. Ginger has
-                        also been used to help treat arthritis, colic, diarrhea, and heart conditions.
-                        It has been used to help treat the common cold, flu-like symptoms, headaches,
-                        and painful menstrual periods.
-                    </p>
-                    <p>
-                        Ginger is native to Asia where it has been used as a cooking spice for at
-                        least 4,400 years.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>Fights germs:</h3>
-                <p>
-                    Certain chemical compounds in fresh ginger help your body ward off germs. They are
-                    especially good at halting growth of bacteria like E.coli and shigella, and they
-                    may also keep viruses like RSV at bay.
-                </p>
-                <h3>Keeps your mouth Healthy:</h3>
-                <p>
-                    Ginger's antibacterial power may also brighten your smile. Active compounds in
-                    ginger called gingerols keep oral bacteria from growing. These bacteria are the
-                    same ones that can cause periodontal disease, a serious gum infection.
-                </p>
-                <h3>Calms nausea:</h3>
-                <p>
-                    The old wives' tale may be true: Ginger helps if you are trying to ease a
-                    queasy stomach, especially during pregnancy. It may work by breaking up
-                    and getting rid of built-up gas in your intestines. It might also help
-                    settle seasickness or nausea caused by chemotherapy.
-                </p>
-                <h3>Soothes sore muscles:</h3>
-                <p>
-                    Ginger will not whisk away muscle pain on the spot, but it may tame soreness
-                    over time. In some studies, people with muscle aches from exercise who
-                    took ginger had less pain the next day than those who did not.
-                </p>
-                <h3>Ease Arthiritis symptoms:</h3>
-                <p>
-                    Ginger is an anti-inflammatory, which means it reduces swelling. That may be
-                    especially helpful for treating symptoms of both rheumatoid arthritis and
-                    osteoarthritis. You might get relief from pain and swelling either by taking
-                    ginger by mouth or by using a ginger compress or patch on your skin.
-                </p>
-                <h3>Curbs cancer growth-Related:</h3>
-                <p>
-                    Some studies show that bioactive molecules in ginger may slow down the
-                    growth of some cancers like colorectal, gastric, ovarian, liver, skin,
-                    breast, and prostate cancer. But much more research is needed to see
-                    if this is true.
-                </p>
-                <h3>Lowers blood sugar:</h3>
-                <p>
-                    One recent small study suggested that ginger may help your body use insulin
-                    better. Larger studies are needed to see if ginger could help improve blood
-                    sugar levels.
-                </p>
-                <h3>Eases period pains:</h3>
-                <p>
-                    Got menstrual cramps? Ginger powder may help. In studies, women who took
-                    1500 milligrams of ginger powder once a day for 3 days during their cycle
-                    felt less pain than women who did not.
-                </p>
-                <h3>Lowers Cholesterol:</h3>
-                <p>
-                    A daily dose of ginger may help you battle your bad or LDL cholesterol
-                    levels. In a recent study, taking 5 grams of ginger a day for 3 months
-                    lowered people's LDL cholesterol an average of 30 points.
-                </p>
-                <h3>Protects against disease:</h3>
-                <p>
-                    Ginger is loaded with antioxidants, compounds that prevent stress and damage
-                    to your body's DNA. They may help your body fight off chronic diseases like
-                    high blood pressure, heart disease, and diseases of the lungs, plus promote
-                    healthy aging.
-                </p>
-                <h3>Relieves Indegestion-Related:</h3>
-                <p>
-                    If you live with chronic indigestion, also called dyspepsia, ginger could
-                    bring some relief. Ginger before meals may make your system empty faster,
-                    leaving less time for food to sit and cause problems.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_SideEffects() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>What are the side effects of using Ginger?</h2>
-                    <p>
-                        Ginger is safe for most people to consume in moderation, but ut is rare
-                        to have side effects from ginger. In high doses it may cause
-                        mild heartburn, abdominal discomfort, diarrhea, and irritation of the mouth.
-                        You may be able to avoid some of the mild stomach side effects, such as
-                        belching, heartburn, or stomach upset, by taking ginger supplements in
-                        capsules or taking ginger with meals.
-                    </p>
-                    <p>
-                        People with gallstones should talk to their doctors before taking ginger.
-                        Be sure to tell your doctor if you are taking ginger before having surgery
-                        or being placed under anesthesia.
-                    </p>
-                    <p>
-                        Pregnant or breastfeeding women, people with heart conditions, and people
-                        with diabetes should not take ginger without talking to their doctors.
-                        Do <span className="em-span">Not</span> take ginger if you have a bleeding
-                        disorder or if you are taking blood-thinning medications, including aspirin.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Dosage() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>What is the recommended dosage of using Ginger?</h2>
-                    <h3>Children:</h3>
-                    <p>
-                        Do <span className="em-span">Not</span> give ginger to children under two.
-                        Children over 2 may take ginger to treat nausea, stomach cramping, and
-                        headaches. Ask your doctor to find the right dose.
-                    </p>
-                    <h3>Adult:</h3>
-                    <p>
-                        Do <span className="em-span">Not</span> take more than 4 g of ginger
-                        per day, including food sources. Pregnant women should not take more
-                        than 1 g per day.
-                    </p>
-                    <p>
-                        For nausea, gas, or indigestion: Some studies have used 1 g of ginger daily,
-                        in divided doses. Ask your doctor to help you find the right dose for you.
-                    </p>
-                    <p>
-                        For pregnancy-induced vomiting: Some studies have used 650 mg to 1 g per
-                        day. Do <span className="em-span">Not</span> take ginger without talking
-                        to your doctor first. For arthritis pain: One study used 250 mg, 4 times
-                        daily.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function GrowIngredient() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Ginger"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="All around the world" />
-                <div className="content-wrapper">
-                    <h2>What do I grow Ginger?</h2>
-
-                    <p>
-                        Because ginger is a tropical plant, it will not live year-round in colder
-                        climates. But that does not mean you cannot grow it at home. Plant it
-                        during the beginning of spring (whenever you're sure there will not be
-                        frost), or keep it potted indoors when temperatures are below 50 degrees
-                        Fahrenheit.
-                    </p>
-                </div>
-            </div>
-            <div>
-                <p>
-                    Growing ginger is pretty simple. Like potatoes, you can start a new ginger plant
-                    from a healthy root you buy at the produce market. Look for one that's about 3 to
-                    5 inches and has some buds growing on it. Soak the root in water for a few hours
-                    before planting to encourage growth.
-                </p>
-                <p>Follow these steps, too:</p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>
-                            Select a pot that's deep enough for the root (a medium-size pot around
-                            15 inches wide).
-                        </li>
-                        <li>
-                            Fill the pot with healthy soil, something full of compost that lets
-                            water in and out.
-                        </li>
-                        <li>Plant the root (buds facing up) 1 inch under the soil.</li>
-                        <li>Place the pot somewhere to get sun for most of the day.</li>
-                        <li>Water regularly to keep the soil slightly moist.</li>
-                        <li>
-                            Harvest the root once the plant is fully grown and the leaves start
-                            to yellow.
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ginger() {
-    return (
-        <>
-            <IngredientNameHeading ingredientName="Ginger" />
-            <div className="Ingredient-hub">
-                <ul>
-                    <li>Overview</li>
-                    <li>How to use</li>
-                    <li>Benefits</li>
-                    <li>Side Effects</li>
-                    <li>Dosage</li>
-                    <li>Grow Ginger</li>
-                </ul>
-            </div>
-            <div className="component-content-rendering">
-                {/* Components will be rendered here using react-router-dom <Link> components*/}
-                <Ingredient_Overview />
-                <Ingredient_HowToUse />
-                <Ingredient_Benefits />
-                <Ingredient_SideEffects />
-                <Ingredient_Dosage />
-                <GrowIngredient />
-            </div >
-        </>
-    )
-}
-
-export default Ginger;
