+++ src/components/layout/Header/Ingredients-content/Spices/Ginger/Ginger.tsx
@@ -1,0 +1,38 @@
+import { Link, Outlet } from "react-router-dom";
+import "../../../../../../App.css";
+import IngredientNameHeading from "../../Ingredient-heading";
+
+function Ginger() {
+    return (
+        <>
+            <IngredientNameHeading ingredientName="Ginger" />
+            <div className="ingredient-hub">
+                <ul>
+                    <li>
+                        <Link to="/Ginger_Overview">Overview</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_HowToUse">How to use</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_Benefits">Benefits</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_SideEffects">Side Effects</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_Dosage">Dosage</Link>
+                    </li>
+                    <li>
+                        <Link to="/Grow_Ginger">Grow Ginger</Link>
+                    </li>
+                </ul>
+            </div>
+            <div className="component-content-rendering">
+                <Outlet />
+            </div >
+        </>
+    )
+}
+
+export default Ginger;
