+++ src/components/layout/Header/Ingredients-content/Spices/Ginger/Ginger_HowToUse.tsx
@@ -1,0 +1,62 @@
+import { Link, Outlet } from "react-router-dom";
+import "../../../../../../App.css";
+import Ingredient from "../../Ingredient";
+import IngredientNameHeading from "../../Ingredient-heading";
+
+function Ginger_HowToUse() {
+    return (
+        <>
+            <IngredientNameHeading ingredientName="Ginger" />
+            <div className="ingredient-hub">
+                <ul>
+                    <li>
+                        <Link to="/Ginger_Overview">Overview</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_HowToUse">How to use</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_Benefits">Benefits</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_SideEffects">Side Effects</Link>
+                    </li>
+                    <li>
+                        <Link to="/Ginger_Dosage">Dosage</Link>
+                    </li>
+                    <li>
+                        <Link to="/Grow_Ginger">Grow Ginger</Link>
+                    </li>
+                </ul>
+            </div>
+            <div className="component-content-rendering">
+                <Outlet />
+                <div className="flex-content">
+                    <Ingredient ingredientName="Ginger"
+                        familyName="Zingiberaceae" origin="Southeast Asia"
+                        geoLocation="All around the world" />
+                    <div className="content-wrapper">
+                        <h2>How is Ginger used?</h2>
+                        <p>
+                            Ginger has been used for thousands of years, even now
+                            where it was used for:
+                        </p>
+                        <div>
+                            <ul>
+                                <li>Foods, Beverages, and Desserts</li>
+                                <li>Beauty, and Cosmetics</li>
+                                <li>Food decoration</li>
+                                <li>Aromatherapy</li>
+                                <li>Medicine which is the mostpopular use in the
+                                    far east such as Asia.
+                                </li>
+                            </ul>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </>
+    )
+}
+
+export default Ginger_HowToUse;
