--- src/components/layout/Header/Ingredients-content/Spices/Turmeric.tsx
@@ -1,646 +1,0 @@
-import "../../../../../App.css"
-import Ingredient from "../Ingredient";
-import IngredientNameHeading from "../Ingredient-heading";
-
-function Ingredient_Overview() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="India, Thailand, Asia" />
-                <div className="content-wrapper">
-                    <h2>What is Turmeric?</h2>
-                    <p>
-                        Turmeric or Curcuma longa is a flowering plant in the ginger family
-                        Zingiberaceae. It is a perennial, rhizomatous, herbaceous plant native
-                        to the Indian subcontinent and Southeast Asia that requires temperatures
-                        between 20 and 30 °C (68 and 86 °F) and high annual rainfall to thrive.
-                    </p>
-                    <p>
-                        Plants are gathered each year for their rhizomes, some for propagation
-                        in the following season and some for consumption.
-                    </p>
-                    <p>
-                        The rhizomes are used fresh or boiled in water and dried, after which
-                        they are ground into a deep orange-yellow powder commonly used as a
-                        coloring and flavoring agent in many Asian cuisines, especially for
-                        curries, as well as for the dyeing characteristics imparted by the
-                        principal turmeric constituent, curcumin.
-                    </p>
-                    <p>
-                        Turmeric powder has a warm, bitter, black pepper-like flavor and earthy,
-                        mustard-like aroma.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_HowToUse() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="Indian subcontinent and southeast Asia" />
-                <div className="content-wrapper">
-                    <h2>How is Turmeric used?</h2>
-                    <p>
-                        Turmeric is a spice used in cooking in many parts of the world.
-                        In India, it is used in curries, in Japan it is served in tea,
-                        and in the U.S. it is used as a supplement and a spice. The
-                        curcumin in turmeric is also available in creams, energy drinks,
-                        cosmetics, and soap.
-                    </p>
-                    <p>
-                        The curcumin in turmeric is poorly absorbed by the human body unless
-                        black pepper is added to it. If using turmeric in cooking, try
-                        adding black pepper to get the most health benefits.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <p>
-                    Turmeric is used in:
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>Cooking</li>
-                        <li>Medicine</li>
-                    </ul>
-                </div>
-                <h3>How it is used:</h3>
-                <p>
-                    Fresh or dried, turmeric has an earthy, pleasantly bitter taste that
-                    magnifies the flavors of other spices. Its peppery, citrusy notes
-                    also make it a delicious spice for seasoning dishes all on its own,
-                    though it's wise to always add a sprinkling of black pepper, which
-                    provides a spicy kick that counterbalances turmeric's astringent side.
-                </p>
-                <div className="ingredient-benefits">
-                    <ul>
-                        <li>
-                            <h3>Ground/Dried Turmeric:</h3>
-                            <dt>
-                                Ground turmeric can also be used to add appetizing color to a breakfast
-                                scramble, a creamy batch of mac and cheese, and even cakes and desserts.
-
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Fresh Turmeric:</h3>
-                            <dt>
-                                If you are lucky enough to come across fresh turmeric root, grab a few
-                                knobby pieces and keep them on hand. Then it's just a matter of grating
-                                the root (the way you woulld grate fresh ginger) and adding it to dishes
-                                where you want turmeric's unique flavors to shine. Fresh turmeric will
-                                stain your hands, so use kitchen gloves when peeling and grating it.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Substituting fresh for dried:</h3>
-                            <dt>
-                                To substitute grated fresh turmeric for dried/ground turmeric, count 1
-                                tablespoon grated fresh turmeric per 1 teaspoon dried/ground.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>How to make Turmeric Tea:</h3>
-                            <dt>
-                                For a hot beverage that is a healthy indulgence, you cannot beat a
-                                steaming cup of turmeric tea. In a mug stir ½ teaspoon of ground
-                                turmeric into ½ to 1 teaspoon of pure maple syrup. Add 1 cup of
-                                boiling water. Add a few grinds of black pepper, and stir to
-                                combine. Doctor the drink up the way you would any cup of tea,
-                                with a squeeze of lemon or a splash of plant milk.
-                            </dt>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Benefits() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="Indian subcontinent and southeast Asia" />
-                <div className="content-wrapper">
-                    <h2>What are the benefits of using Turmeric?</h2>
-                    <p>
-                        Traditional claims that turmeric contains compounds with medicinal
-                        properties. These compounds are called curcuminoids. The most important
-                        one is curcumin. Curcumin is the main active ingredient in turmeric.
-                        It has powerful anti-inflammatory effects and is a very strong antioxidant:
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>Benefits of using Turmeric:</h3>
-                <div className="ingredient-benefits">
-                    <ol>
-                        <li>
-                            <h3>
-                                Turmeric contains bioactive compounds with
-                                medicinal properties:
-                            </h3>
-                            <dt>
-                                The curcumin content of turmeric isn't that high. it is around 3%,
-                                by weight. Most of the studies on this herb use
-                                turmeric extracts that contain mostly curcumin itself, with
-                                dosages usually exceeding 1 gram per day.
-                            </dt>
-                            <dt>
-                                It would be very difficult to reach these levels just using
-                                turmeric as a spice in your foods. That's why some people choose
-                                to use supplements.
-                            </dt >
-                            <dt>
-                                Curcumin, however, is poorly absorbed into your bloodstream. In
-                                order to experience the full effects of curcumin, its
-                                bioavailability (the rate at which your body absorbs a substance)
-                                needs to improve.
-
-                                It helps to consume it with black pepper, which contains piperine.
-                                Piperine is a natural substance that enhances the absorption of
-                                curcumin by <a href="https://pubmed.ncbi.nlm.nih.gov/9619120/">
-                                    2000%</a>.
-                            </dt>
-                            <dt>
-                                In fact, the best curcumin supplements contain piperine, and this
-                                makes them substantially more effective. Curcumin is also fat
-                                soluble, which means it breaks down and dissolves in fat or oil.
-                                That's why it may be a good idea to take curcumin supplements
-                                with a meal that's high in fat.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Curcumin is a natural anti-inflammatory compound:</h3>
-                            <dt>
-                                Inflammation is incredibly important. It helps fight foreign
-                                invaders and has a role in repairing damage in your body.
-                            </dt>
-                            <dt>
-                                Although acute, short-term inflammation is beneficial, it can
-                                be a concern if it becomes chronic and attacks your body's own
-                                tissues.
-                            </dt>
-                            <dt>
-                                Scientists now believe that chronic low-level inflammation can
-                                play a role in some health conditions and diseases.
-                                These include:
-                            </dt>
-                            <ul>
-                                <li>Heart disease</li>
-                                <li>Cancer</li>
-                                <li>Metabolic syndrome</li>
-                                <li>Alzheimer's disease</li>
-                                <li>Various degenerative conditions</li>
-                            </ul>
-                            <dt>
-                                That's why anything that can help fight chronic inflammation is
-                                potentially important in preventing and helping treat these
-                                conditions.
-                            </dt>
-                            <dt>
-                                While the topic of inflammation is multilayered and there's
-                                likely no simple answer, the key takeaway regarding curcumin
-                                is that it is a bioactive substance that can fight inflammation.
-                                However, very high doses are required to produce
-                                medicinal results.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>
-                                Turmeric can increase the antioxidant capacity of the body:
-                            </h3>
-                            <dt>
-                                Oxidative damage is believed to be one of the mechanisms behind
-                                aging and many diseases.
-                            </dt>
-                            <dt>
-                                It involves free radicals, highly reactive molecules with unpaired
-                                electrons. Free radicals tend to react with important organic
-                                substances, such as fatty acids, proteins, or DNA.
-                            </dt>
-                            <dt>
-                                The main reason antioxidants are so beneficial is that they protect
-                                your body from free radicals. Curcumin is a potent antioxidant
-                                that can neutralize free radicals due to its chemical structure.
-                            </dt>
-                            <dt>
-                                In addition, animal and cellular studies suggest that
-                                curcumin may block the action of free radicals and may stimulate
-                                the action of other antioxidants. Further clinical studies are
-                                needed in humans to confirm these benefits.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>
-                                Curcumin can boost brain-derived neurotrophic factor:
-                            </h3>
-                            <dt>
-                                Before scientists had a better understanding of neurons, it was
-                                believed that they weren't able to divide and multiply after
-                                early childhood. However, they now know that isn't the case.
-                                <dt>
-                                    Neurons are capable of forming new connections, and in certain
-                                    areas of the brain they can multiply and increase in number.
-
-                                    One of the main drivers of this process is brain-derived
-                                    neurotrophic factor (BDNF). This is a gene that's involved in
-                                    making a protein responsible for promoting the life of neurons.
-                                </dt>
-                                <dt>
-                                    The BDNF protein plays a role in memory and learning, and it can
-                                    be found in areas of the brain responsible for eating, drinking,
-                                    and body weight.
-
-                                    Many common brain disorders have been linked to decreased levels
-                                    of BDNF protein, including depression and
-                                    Alzheimer's disease.
-                                </dt>
-                                <dt>
-                                    Interestingly, animal studies have found that curcumin may increase
-                                    brain levels of BDNF. By doing this, it may be effective in
-                                    delaying or even reversing many brain diseases and age-related
-                                    decreases in brain function. Still, since these studies were
-                                    performed in animals, it is hard to say what the results mean
-                                    for humans.
-                                </dt>
-                                <dt>
-                                    It may also help improve memory and attention, which seems logical
-                                    given its effects on BDNF levels. However, more studies are needed
-                                    to confirm this.
-                                </dt>
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Curcumin may lower your risk of heart disease:</h3>
-                            <dt>
-                                Heart disease is the number one cause of death in the world.
-                                Researchers have studied it for many decades and learned a lot
-                                about why it happens. Unsurprisingly, heart disease is incredibly
-                                complicated and various things contribute to it.
-                            </dt>
-                            <dt>
-                                Curcumin may help reverse many steps in the heart disease
-                                process. Perhaps the main benefit of curcumin when it comes to
-                                heart disease is improving the function of the endothelium, the
-                                lining of your blood vessels.
-                            </dt>
-                            <dt>
-                                The endothelial dysfunction is a major driver of heart disease.
-                                This is when your endothelium is unable to regulate blood pressure,
-                                blood clotting, and various other factors.
-
-                            </dt>
-                            <dt>
-                                Several studies suggest that curcumin can lead to improvements
-                                in heart health. Additionally, one study found that it is as
-                                effective as exercise in post-menopausal women.
-                            </dt>
-                            <dt>
-                                In addition, curcumin can help reduce inflammation and oxidation
-                                (as discussed above), which can play a role in heart disease.
-                                In one study of 121 people undergoing coronary artery bypass
-                                surgery, researchers assigned them either a placebo or 4 grams of
-                                curcumin per day a few days before and after the surgery.
-                            </dt>
-                            <dt>
-                                The curcumin group had a 65% decreased risk of experiencing a
-                                heart attack in the hospital.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>
-                                Turmeric may help prevent cancer:
-                            </h3>
-                            <dt>
-                                Cancer is a disease, characterized by uncontrolled cell growth.
-                                There are many different forms of cancer that appear to be affected
-                                by curcumin supplements.
-                            </dt>
-                            <dt>
-                                Curcumin has been studied as a beneficial herb in cancer
-                                treatment and been found to affect cancer
-                                growth and development.
-                            </dt>
-                            <dt>
-                                Studies have shown that it can:
-                            </dt>
-                            <ul>
-                                <li>
-                                    Contribute to the death of cancerous cells.
-                                </li>
-                                <li>
-                                    Reduce angiogenesis (growth of new blood vessels in tumors)
-                                    Reduce metastasis (spread of cancer)
-                                </li>
-                                <li>
-                                    Whether high-dose curcumin — preferably with an absorption
-                                    enhancer like piperine - can help treat cancer in humans has
-                                    yet to be studied properly.
-                                </li>
-                            </ul>
-                            <dt>
-                                However, there is evidence that it may prevent cancer
-                                from occurring in the first place, especially cancers of the digestive
-                                system like colorectal cancer. In a 30-day study in
-                                44 men with lesions in the colon that sometimes turn cancerous,
-                                4 grams of curcumin per day reduced the number of lesions by 40%.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>
-                                Curcumin may be useful in treating Alzheimer's disease:
-                            </h3>
-                            <dt>
-                                Alzheimer's disease is the most common form of dementia and may
-                                contribute to up 70% of dementia cases.
-                                While treatment exits for some of its symptoms, there is no
-                                cure for Alzheimer's yet. That's why preventing it from occurring
-                                in the first place is so important.
-                            </dt>
-                            <dt>
-                                There may be good news on the horizon because curcumin has been
-                                shown to cross the blood-brain barrier.
-                            </dt>
-                            <dt>
-                                it is known that inflammation and oxidative damage play a role in
-                                Alzheimer's disease, and curcumin has beneficial effectsTrusted
-                                Source on both.
-                            </dt>
-                            <dt>
-                                In addition, a key feature of Alzheimer's disease is a buildup of
-                                protein tangles called amyloid plaques. Studies show that curcumin
-                                can help clear these plaques.
-                            </dt>
-                            <dt>
-                                Whether curcumin can slow or even reverse the progression of
-                                Alzheimer's disease in people is currently unknown and needs to be
-                                studied.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Arthritis patients respond well to curcumin supplements:</h3>
-                            <dt>
-                                Arthritis is a common problem in Western countries. There are several
-                                different types of arthritis, most of which involve inflammation in
-                                the joints.
-                            </dt>
-                            <dt>
-                                Given that curcumin is a potent anti-inflammatory compound, it makes
-                                sense that it may help with arthritis. In fact, several studies show
-                                that there is an association.
-                            </dt>
-                            <dt>
-                                In a study in people with rheumatoid arthritis,
-                                curcumin was even more effective than an anti-inflammatory drug.
-
-                                May other studies have looked at the effects of
-                                curcumin on arthritis and noted improvements in various symptoms.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Curcumin has benefits against depression:</h3>
-                            <dt>
-                                Curcumin has shown some promise in treating depression.
-                                In a controlled trial, 60 people with depression
-                                were randomized into three groups. One group took Prozac, another
-                                group took 1 gram of curcumin, and the third group took both
-                                Prozac and curcumin.
-                            </dt>
-                            <dt>
-                                After 6 weeks, curcumin had led to improvements similar to those
-                                of Prozac. The group that took both Prozac and curcumin fared best.
-                                According to this small study, curcumin is as effective as an
-                                antidepressant.
-                            </dt>
-                            <dt>
-                                Depression is also linked to reduced levels of BDNF and a shrinking
-                                hippocampus, a brain area with a role in learning and memory.
-                                Curcumin can help boost BDNF levels, potentially reversing some of
-                                these changes.
-                            </dt>
-                            <dt>
-                                There's also some evidence that curcumin can boost the brain
-                                neurotransmitters serotonin and dopamine.
-                            </dt>
-                        </li>
-                        <li>
-                            <h3>Curcumin may help delay aging and fight age-related
-                                chronic diseases:
-                            </h3>
-                            <dt>
-                                If curcumin can really help prevent heart disease, cancer, and
-                                Alzheimer's, it may have benefits for longevity as well.
-                                This suggests that curcumin may have potential as an anti-aging
-                                supplement.
-                            </dt>
-                            <dt>
-                                Given that oxidation and inflammation are believed to play a
-                                role in aging, curcumin may have effects that go way beyond just
-                                preventing disease.
-                            </dt>
-                        </li>
-                    </ol>
-                </div >
-            </div >
-        </>
-    )
-}
-
-function Ingredient_SideEffects() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="Indian subcontinent and southeast Asia" />
-                <div className="content-wrapper">
-                    <h2>What are the side effects of using Turmeric?</h2>
-                    <p>
-                        It is possible to take too much turmeric, but it is not
-                        likely. One study found that heavy doses may be toxic
-                        to cells, but other research shows that supplements of
-                        less than 8,000mg daily is safe. More research is
-                        needed to find the best dose for turmeric.
-                    </p>
-                    <p>
-                        Turmeric is complex and difficult to study because it
-                        easily changes in your body and little reaches your
-                        bloodstream. Because of that, some manufacturers have
-                        produced turmeric or curcumin products that have better
-                        absorption rates.
-                    </p>
-                    <p>
-                        It is important to read the label clearly and speak with
-                        your healthcare provider about the amount of turmeric
-                        you take, if you choose to take it.
-                    </p>
-                    <p>The side effects may include:</p>
-                    <div className="ingredient-benefits">
-                        <ul>
-                            <li>Skin rash</li>
-                            <li>Diarrhea or other gastrointestinal discomfort</li>
-                            <li>Nausea</li>
-                            <li>Headache</li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function Ingredient_Dosage() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="Indian subcontinent and southeast Asia" />
-                <div className="content-wrapper">
-                    <h2>What is the recommended dosage for using Turmeric?</h2>
-                    <p>
-                        Some companies are developing turmeric or curcumin products
-                        that have increased bioavailability so they are absorbed better. It's
-                        important to remember that herbs that are reformulated to be absorbed
-                        better may also increase the harmful effects of the supplement too.
-                        The Food and Drug Administration (FDA) has approved curcuminoids—anything
-                        isolated from turmeric—and calls them generally recognized as safe.
-                        Supplements from turmeric with 4,000 to 8,000 milligrams per day are
-                        recognized as safe.
-                    </p>
-                </div>
-            </div>
-        </>
-    )
-}
-
-function GrowIngredient() {
-    return (
-        <>
-            <div className="flex-content">
-                <Ingredient ingredientName="Turmeric"
-                    familyName="Zingiberaceae" origin="Southeast Asia"
-                    geoLocation="Indian subcontinent and southeast Asia" />
-                <div className="content-wrapper">
-                    <h2>How do I grow Turmeric?</h2>
-                    <p>
-                        One of the great joys of gardening (and among the most satisfying) is tasting
-                        the fruits of our labor, especially when those labors include adding herbs
-                        and spices to the garden mix. Even in climates such as ours with its short
-                        growing season and freezing winter weather, it's possible to grow tropical
-                        spices like ginger and turmeric (U.S. hardiness zones 8-10). You just need
-                        to grow them indoors.
-                    </p>
-                </div>
-            </div>
-            <div className="content-section2">
-                <h3>Growing Turmeric:</h3>
-                <p>
-                    Two spices, ginger (Zingiber officinale) and turmeric (Curcuma longa), are
-                    easy to grow in containers and adapt well to living inside although they do
-                    benefit from time outside during the warm weather months. Indoors or out,
-                    both grow best at temperatures of at least 68 degrees Fahrenheit and should
-                    not be exposed to temperatures below 50.
-                </p>
-                <p>
-
-                    Both ginger and turmeric grow from a rhizome (a fleshy root) that is the
-                    edible part of the plant. The rhizomes often can be purchased through a
-                    mail-order nursery or at a grocery store. Try an organic market if you cannot
-                    find these in your local supermarket.
-                </p>
-                <p>
-                    Select one that is firm (not shriveled), several inches long and with several
-                    buds or "eyes." Cut it into 2-3-inch pieces, making sure that each piece
-                    includes at least 2-3 buds.
-                </p>
-                <p>
-                    Choose a pot at least 12 inches in diameter. Fill the pot almost to the top
-                    with a nutrient-rich soil that will drain well. Place the rhizome
-                    horizontally on the surface with the eyes facing up. Cover with about 2
-                    inches of soil.
-                </p>
-                <p>
-                    Water sparingly until sprouts appear. The soil should be moist, but never wet.
-                    Too much moisture may rot the rhizome.Put the pot in a warm location (70-80 degrees
-                    Fahrenheit). If the room is cool, a heat mat such as those used for seed starting
-                    can be used to warmthe soil. Remove from the heat mat once you see sprouts.
-                </p>
-                <p>
-                    Now comes the most difficult part. Patience. It will take 3-8 weeks for
-                    sprouts to appear. Ginger will likely sprout sooner, turmeric later. When
-                    they do, move the pot to a sunny spot near a window. If necessary, supplement
-                    with a grow light. Water as needed to keep the soil moist (not wet), and mist
-                    often if the air is dry.
-                </p>
-                <p>
-                    For the most part, ginger and turmeric plants are relatively care free with
-                    ginger growing 2-3 feet in height and turmeric 3-4 feet. As they mature,
-                    you'll get a lovely display of tropical foliage and perhaps a flower or two.
-                </p>
-                <p>
-                    When the weather gets warmer, and after your plants have begun to leaf out,
-                    you can move your ginger and turmeric plants outside to benefit from the
-                    fresh air and sunshine. Daytime temperatures should be at least 68 degrees
-                    Fahrenheit, with nighttime temperatures not below 50.
-                </p>
-                <p>
-                    Gradually introduce the plants to sunlight over the course of several days
-                    to avoid burning the foliage. When the weather begins to turn colder, or
-                    temperatures drop below 50, bring them back inside.
-                </p>
-                <p>
-                    After 8-10 months, the foliage will begin to fade. When the plant has
-                    yellowed and dried out, it's time to harvest. Simply dig up the entire plant,
-                    brush the soil away from the rhizomes, and cut the stalk off.
-                </p>
-                <p>
-                    Save a piece or two of the rhizomes to replant. Rinse off the rest and
-                    allow to dry. The whole rhizome can be stored in the refrigerator for up
-                    to three months, or cut in pieces and stored in the freezer for up to six
-                    months.
-                </p>
-            </div>
-        </>
-    )
-}
-
-function Turmeric() {
-    return (
-        <>
-            <IngredientNameHeading ingredientName="Turmeric" />
-            <div className="Ingredient-hub">
-                <ul>
-                    <li>Overview</li>
-                    <li>How to use</li>
-                    <li>Benefits</li>
-                    <li>Side Effects</li>
-                    <li>Dosage</li>
-                    <li>Grow Turmeric</li>
-                </ul>
-            </div>
-            <div className="component-content-rendering">
-                {/* Components will be rendered here using react-router-dom <Link> components*/}
-                <Ingredient_Overview />
-                <Ingredient_HowToUse />
-                <Ingredient_Benefits />
-                <Ingredient_SideEffects />
-                <Ingredient_Dosage />
-                <GrowIngredient />
-            </div>
-        </>
-    )
-}
-
-export default Turmeric;
