+++ src/components/layout/Header/Ingredients-content/Spices/Turmeric/Turmeric.tsx
@@ -1,0 +1,38 @@
+import { Link, Outlet } from "react-router-dom";
+import "../../../../../../App.css";
+import IngredientNameHeading from "../../Ingredient-heading";
+
+function Turmeric() {
+    return (
+        <>
+            <IngredientNameHeading ingredientName="Turmeric" />
+            <div className="ingredient-hub">
+                <ul>
+                    <li>
+                        <Link to="/Turmeric_Overview">Overview</Link>
+                    </li>
+                    <li>
+                        <Link to="/Turmeric_HowToUse">How to use</Link>
+                    </li>
+                    <li>
+                        <Link to="/Turmeric_Benefits">Benefits</Link>
+                    </li>
+                    <li>
+                        <Link to="/Turmeric_SideEffects">Side Effects</Link>
+                    </li>
+                    <li>
+                        <Link to="/Turmeric_Dosage">Dosage</Link>
+                    </li>
+                    <li>
+                        <Link to="/Grow_Turmeric">Grow Turmeric</Link>
+                    </li>
+                </ul>
+            </div>
+            <div className="component-content-rendering">
+                <Outlet />
+            </div>
+        </>
+    )
+}
+
+export default Turmeric;
